osmo-bts-omldummy: Fix crash accessing null phy
osmo-bts-omldummy is known to be specially in the sense that it has no
phy. This should ideally be changed to have some dummy object, but so
far that's how it is. A recent commit checking the phy's forgot about
the fact that BTS model has no phy. Let's fix it.

Fixes: 4ddc37ce7140e5ce6ca6cce5aeeefedc0d395d22
Change-Id: Iba7873530aac74b3bf799819890641196dacf170